+++ dj_anonymizer/anonymizer.py
@@ -1,8 +1,17 @@
-from django.apps import apps
-from django_bulk_update.helper import bulk_update
+import django
 
 from dj_anonymizer.conf import settings
 from dj_anonymizer.utils import import_if_exist
+
+
+if django.__version__ <= '2.2':
+    try:
+        from django_bulk_update.helper import bulk_update
+    except ModuleNotFoundError:
+        raise ModuleNotFoundError(
+            "Django %s does not have native support for bulk_update and "
+            "django_bulk_update is not installed""" % django.__version__
+        )
 
 
 class Anonymizer:
@@ -17,7 +26,7 @@ class Anonymizer:
         # as single file for defining all models to anonymize
         import_if_exist('base')
 
-        for app in apps.get_app_configs():
+        for app in django.apps.apps.get_app_configs():
             models_set.update(
                 model.__module__ + '.' + model.__name__
                 for model in app.get_models()
@@ -54,19 +63,27 @@ class Anonymizer:
             total = queryset.count()
             for j in list(range(0, total,
                           settings.ANONYMIZER_SELECT_BATCH_SIZE)) + [None]:
-                sub_set = queryset.order_by('pk')[i:j]
-                for model in sub_set:
+                subset = queryset.order_by('pk')[i:j]
+                for obj in subset:
                     i += 1
 
                     for name in anonym_cls.get_fields_names():
-                        if getattr(model, name) or anonym_cls.Meta.fill_empty:
-                            setattr(model, name, next(
-                                getattr(anonym_cls, name))
-                            )
+                        setattr(obj, name, next(
+                            getattr(anonym_cls, name))
+                        )
 
-                bulk_update(sub_set,
-                            batch_size=settings.ANONYMIZER_UPDATE_BATCH_SIZE,
-                            update_fields=anonym_cls.get_fields_names())
+                if django.__version__ <= '2.2':
+                    bulk_update(
+                        subset,
+                        batch_size=settings.ANONYMIZER_UPDATE_BATCH_SIZE,
+                        update_fields=anonym_cls.get_fields_names()
+                    )
+                else:
+                    subset.model.objects.bulk_update(
+                        subset,
+                        anonym_cls.get_fields_names(),
+                        batch_size=settings.ANONYMIZER_UPDATE_BATCH_SIZE,
+                    )
         print('\n\nUpdating finished')
 
     def clean(self):